Swap module: remove old swap classes
- extract fetching coins list into CoinSelect module
