+++ UnstoppableWallet/UnstoppableWallet/Modules/Swap/CoinCard/CoinCardModule.swift
@@ -1,0 +1,13 @@
+struct CoinCardModule {
+
+    static func fromCell(service: SwapService, tradeService: SwapTradeService) -> SwapCoinCardCell {
+        let viewModel = SwapFromCoinCardViewModel(service: service, tradeService: tradeService, decimalParser: AmountDecimalParser())
+        return SwapCoinCardCell(viewModel: viewModel)
+    }
+
+    static func toCell(service: SwapService, tradeService: SwapTradeService) -> SwapCoinCardCell {
+        let viewModel = SwapToCoinCardViewModel(service: service, tradeService: tradeService, decimalParser: AmountDecimalParser())
+        return SwapCoinCardCell(viewModel: viewModel)
+    }
+
+}
