+++ UnstoppableWallet/UnstoppableWallet/Modules/Swap/SwapService.swift
@@ -1,536 +1,245 @@
 import RxSwift
-import RxCocoa
 import RxRelay
 import HsToolKit
 import UniswapKit
 import CurrencyKit
 import BigInt
+import EthereumKit
+import Foundation
 
 class SwapService {
-    private static let refreshInterval: TimeInterval = 10
-    private static let warningPriceImpact: Decimal = 1
-    private static let forbiddenPriceImpact: Decimal = 5
-
-    private let disposeBag = DisposeBag()
-    private var allowanceDisposable: Disposable?
-    private var tradeDataDisposable: Disposable?
-    private var swapDisposable: Disposable?
-    private var feeDisposable: Disposable?
-
-    private let uniswapRepository: UniswapRepository
-    private let allowanceProvider: AllowanceProvider
-    private let swapFeeRepository: SwapFeeRepository
-    private let swapCoinProvider: SwapCoinProvider
+    private let ethereumKit: EthereumKit.Kit
+    private let tradeService: SwapTradeService
+    private let allowanceService: SwapAllowanceService
+    private let pendingAllowanceService: SwapPendingAllowanceService
+    private let transactionService: EthereumTransactionService
     private let adapterManager: IAdapterManager
 
-    private var estimatedRelay = PublishRelay<TradeType>()
-    private var coinInRelay = PublishRelay<Coin?>()
-    private var coinOutRelay = PublishRelay<Coin?>()
+    private let disposeBag = DisposeBag()
 
-    private var amountInRelay = PublishRelay<Decimal?>()
-    private var amountOutRelay = PublishRelay<Decimal?>()
-
-    private var balanceInRelay = BehaviorRelay<Decimal?>(value: nil)
-    private var balanceOutRelay = BehaviorRelay<Decimal?>(value: nil)
-    private var validationErrorsRelay = BehaviorRelay<[Error]>(value: [])
-
-    private var tradeDataStateRelay = PublishRelay<DataStatus<SwapModule.TradeItem>?>()
-    private var allowanceStateRelay = BehaviorRelay<DataStatus<Decimal>?>(value: nil)
-    private var feeStateRelay = PublishRelay<DataStatus<SwapModule.SwapFeeInfo>?>()
-
-    private var swapStateRelay = BehaviorRelay<DataStatus<Data>?>(value: nil)
-    private var stateRelay = BehaviorRelay<SwapModule.SwapState>(value: .idle)
-
-    private var tradeData: TradeData?
-    var tradeOptions: TradeOptions = TradeOptions()
-
-    private var approvingTimer: Timer?
-    private var lastAllowance: Decimal?
-
-    private(set) var estimated = TradeType.exactIn {
+    private let stateRelay = PublishRelay<State>()
+    private(set) var state: State = .notReady {
         didSet {
-            estimatedRelay.accept(estimated)
+            if oldValue != state {
+                stateRelay.accept(state)
+            }
         }
     }
 
-    private(set) var coinIn: Coin? {
-        didSet {
-            coinInRelay.accept(coinIn)
+    private let swapEventRelay = PublishRelay<SwapEvent>()
 
-            lastAllowance = nil
-            updateAllowance()
+    private let errorsRelay = PublishRelay<[Error]>()
+    private(set) var errors: [Error] = [] {
+        didSet {
+            errorsRelay.accept(errors)
         }
     }
 
-    private(set) var coinOut: Coin? {
-        didSet {
-            coinOutRelay.accept(coinOut)
-        }
-    }
-
-    private(set) var amountIn: Decimal? {
-        didSet {
-            amountInRelay.accept(amountIn)
-        }
-    }
-
-    private(set) var amountOut: Decimal? {
-        didSet {
-            amountOutRelay.accept(amountOut)
-        }
-    }
-
+    private let balanceInRelay = PublishRelay<Decimal?>()
     private(set) var balanceIn: Decimal? {
         didSet {
             balanceInRelay.accept(balanceIn)
         }
     }
 
+    private let balanceOutRelay = PublishRelay<Decimal?>()
     private(set) var balanceOut: Decimal? {
         didSet {
             balanceOutRelay.accept(balanceOut)
         }
     }
 
-    private(set) var tradeDataState: DataStatus<SwapModule.TradeItem>? {
-        didSet {
-            tradeDataStateRelay.accept(tradeDataState)
-        }
-    }
-
-    private(set) var feeState: DataStatus<SwapModule.SwapFeeInfo>? {
-        didSet {
-            feeStateRelay.accept(feeState)
-        }
-    }
-
-    var feePriority: FeeRatePriority { swapFeeRepository.priority }
-
-    init(uniswapRepository: UniswapRepository, allowanceRepository: AllowanceProvider, swapFeeRepository: SwapFeeRepository, swapCoinProvider: SwapCoinProvider, adapterManager: IAdapterManager, coin: Coin) {
-        self.uniswapRepository = uniswapRepository
-        self.allowanceProvider = allowanceRepository
-        self.swapFeeRepository = swapFeeRepository
-        self.swapCoinProvider = swapCoinProvider
+    init(ethereumKit: EthereumKit.Kit, tradeService: SwapTradeService, allowanceService: SwapAllowanceService, pendingAllowanceService: SwapPendingAllowanceService, transactionService: EthereumTransactionService, adapterManager: IAdapterManager) {
+        self.ethereumKit = ethereumKit
+        self.tradeService = tradeService
+        self.allowanceService = allowanceService
+        self.pendingAllowanceService = pendingAllowanceService
+        self.transactionService = transactionService
         self.adapterManager = adapterManager
 
-        coinIn = coin
+        tradeService.stateObservable
+                .subscribeOn(ConcurrentDispatchQueueScheduler(qos: .userInitiated))
+                .subscribe(onNext: { [weak self] state in
+                    self?.onUpdateTrade(state: state)
+                })
+                .disposed(by: disposeBag)
 
-        updateBalances()
-        updateAllowance()
+        tradeService.coinInObservable
+                .subscribeOn(ConcurrentDispatchQueueScheduler(qos: .userInitiated))
+                .subscribe(onNext: { [weak self] coin in
+                    self?.onUpdate(coinIn: coin)
+                })
+                .disposed(by: disposeBag)
+        onUpdate(coinIn: tradeService.coinIn)
 
-        sync()
+        tradeService.coinOutObservable
+                .subscribeOn(ConcurrentDispatchQueueScheduler(qos: .userInitiated))
+                .subscribe(onNext: { [weak self] coin in
+                    self?.onUpdate(coinOut: coin)
+                })
+                .disposed(by: disposeBag)
+
+        tradeService.amountInObservable
+                .subscribeOn(ConcurrentDispatchQueueScheduler(qos: .userInitiated))
+                .subscribe(onNext: { [weak self] amount in
+                    self?.onUpdate(amountIn: amount)
+                })
+                .disposed(by: disposeBag)
+
+        allowanceService.stateObservable
+                .subscribeOn(ConcurrentDispatchQueueScheduler(qos: .userInitiated))
+                .subscribe(onNext: { [weak self] _ in
+                    self?.syncState()
+                })
+                .disposed(by: disposeBag)
+
+        pendingAllowanceService.isPendingObservable
+                .subscribeOn(ConcurrentDispatchQueueScheduler(qos: .userInitiated))
+                .subscribe(onNext: { [weak self] isPending in
+                    self?.onUpdate(isAllowancePending: isPending)
+                })
+                .disposed(by: disposeBag)
+
+        transactionService.transactionStatusObservable
+                .subscribeOn(ConcurrentDispatchQueueScheduler(qos: .userInitiated))
+                .subscribe(onNext: { [weak self] _ in
+                    self?.syncState()
+                })
+                .disposed(by: disposeBag)
     }
 
-    private func startWaitingApprove() {
-        guard approvingTimer == nil else {
-            return
+    private var ethereumBalance: BigUInt {
+        ethereumKit.balance ?? 0
+    }
+
+    private func onUpdateTrade(state: SwapTradeService.State) {
+        if case .ready(let trade) = state {
+            if let kitTransactionData = try? tradeService.transactionData(tradeData: trade.tradeData) { // todo: handle throwing function correctly
+                let transactionData = EthereumTransactionService.TransactionData(
+                        to: kitTransactionData.to,
+                        value: kitTransactionData.value,
+                        input: kitTransactionData.input
+                )
+
+                transactionService.set(transactionData: transactionData)
+            }
+        } else {
+            transactionService.set(transactionData: nil)
         }
 
-        approvingTimer = Timer.scheduledTimer(withTimeInterval: Self.refreshInterval, repeats: true) { [weak self] _ in
-            self?.updateAllowance()
-        }
-
-        updateAllowance()
+        syncState()
     }
 
-    private func stopWaitingApprove() {
-        if approvingTimer != nil {
-            approvingTimer?.invalidate()
-            approvingTimer = nil
-        }
+    private func onUpdate(coinIn: Coin?) {
+        balanceIn = coinIn.flatMap { balance(coin: $0) }
+        allowanceService.set(coin: coinIn)
+        pendingAllowanceService.set(coin: coinIn)
     }
 
-    private func coin(for type: TradeType) -> Coin? {
-        type == .exactIn ? coinIn : coinOut
+    private func onUpdate(amountIn: Decimal?) {
+        syncState()
     }
 
-    private func amount(for type: TradeType) -> Decimal? {
-        type == .exactIn ? amountIn : amountOut
+    private func onUpdate(coinOut: Coin?) {
+        balanceOut = coinOut.flatMap { balance(coin: $0) }
     }
 
-    private func allowanceAmount(for type: TradeType) -> Decimal? {
-        type == .exactIn ? amountIn : tradeData?.amountInMax
-    }
+    private func onUpdate(isAllowancePending: Bool) {
+        syncState()
 
-    private func insufficientAllowance() -> Bool {
-        if let allowance = allowanceStateRelay.value?.data,
-           (allowanceAmount(for: estimated) ?? 0) > allowance {     // not enough allowance for swap
-            return true
-        }
-        return false
-    }
-
-    private func clearEstimated(for type: TradeType) {
-        switch type {
-        case .exactIn:
-            amountOut = nil
-        case .exactOut:
-            amountIn = nil
+        if case .failed = transactionService.transactionStatus, !isAllowancePending {
+            transactionService.resync() // after required allowance is approved, transaction service state should be resynced
         }
     }
 
-    private func balance(coin: Coin?) -> Decimal? {
-        guard let coin = coin, let adapter = adapterManager.adapter(for: coin) as? IBalanceAdapter else {
+    private func syncState() {
+        var allErrors = [Error]()
+        var loading = false
+
+        switch tradeService.state {
+        case .loading:
+            loading = true
+        case .ready(let trade):
+            if let impactLevel = trade.impactLevel, impactLevel == .forbidden {
+                allErrors.append(SwapError.forbiddenPriceImpactLevel)
+            }
+        case .notReady(let errors):
+            allErrors.append(contentsOf: errors)
+        }
+
+        if let allowanceState = allowanceService.state {
+            switch allowanceState {
+            case .loading:
+                loading = true
+            case .ready(let allowance):
+                if let amountIn = tradeService.amountIn, amountIn > allowance {
+                    allErrors.append(SwapError.insufficientAllowance)
+                }
+            case .notReady(let error):
+                allErrors.append(error)
+            }
+        }
+
+        if let amountIn = tradeService.amountIn, let balanceIn = balanceIn, amountIn > balanceIn {
+            allErrors.append(SwapError.insufficientBalanceIn)
+        }
+
+        switch transactionService.transactionStatus {
+        case .loading:
+            loading = true
+        case .completed(let transaction):
+            if transaction.totalAmount > ethereumBalance {
+                allErrors.append(TransactionError.insufficientBalance(requiredBalance: transaction.totalAmount))
+            }
+        case .failed(let error):
+            if !allErrors.contains(where: { error in
+                switch error {
+                case _ as SwapError: return true
+                case _ as TransactionError: return true
+                default: return false
+                }
+            }) {
+                allErrors.append(error)
+            }
+        }
+
+        if pendingAllowanceService.isPending {
+            loading = true
+        }
+
+        errors = allErrors
+
+        if loading {
+            state = .loading
+        } else if allErrors.isEmpty {
+            state = .ready
+        } else {
+            state = .notReady
+        }
+    }
+
+    private func balance(coin: Coin) -> Decimal? {
+        guard let adapter = adapterManager.adapter(for: coin) as? IBalanceAdapter else {
             return nil
         }
 
         return adapter.balance
     }
 
-    private func updateFeeState() {
-        guard let coinIn = coinIn, let tradeData = tradeData else {
-            feeState = nil
-            return
-        }
-
-        if allowanceStateRelay.value != nil, insufficientAllowance() {    // for erc20 without allowance(or loading) can't calculate fee
-            feeState = nil
-            return
-        }
-
-        feeDisposable?.dispose()
-
-        feeState = .loading
-        feeDisposable = swapFeeRepository.swapFeeInfo(coin: coinIn, tradeData: tradeData)
-                .subscribe(onSuccess: { [weak self] info in
-                    self?.feeState = .completed(info)
-                    self?.sync()
-                }, onError: { [weak self] error in
-                    self?.feeState = .failed(error)
-                    self?.sync()
-                })
-        feeDisposable?.disposed(by: disposeBag)
-    }
-
-    private func updateTradeData() {
-        guard let coinIn = coinIn, let coinOut = coinOut else {
-
-            tradeDataState = nil
-            return
-        }
-
-        let amount = self.amount(for: estimated) ?? 0
-
-        tradeDataDisposable?.dispose()
-
-        tradeDataState = .loading
-        tradeDataDisposable = uniswapRepository
-                .trade(coinIn: coinIn, coinOut: coinOut, amount: amount, tradeType: estimated, tradeOptions: TradeOptions())
-                .subscribe(onSuccess: { [weak self] item in
-                    self?.handle(tradeData: item, coinIn: coinIn, coinOut: coinOut)
-                }, onError: { [weak self] error in
-                    self?.tradeDataState = .failed(error)
-                    self?.sync()
-                })
-
-        tradeDataDisposable?.disposed(by: disposeBag)
-    }
-
-    private func updateAllowance() {
-        allowanceDisposable?.dispose()
-
-        guard let coinIn = self.coinIn, coinIn.type != .ethereum else {
-            allowanceStateRelay.accept(nil)
-
-            return
-        }
-
-        allowanceStateRelay.accept(.loading)
-        allowanceDisposable = allowanceProvider
-                .allowanceObservable(coin: coinIn, spenderAddress: uniswapRepository.routerAddress)
-                .subscribe(onSuccess: { [weak self] allowance in
-                    self?.handle(allowance: allowance, coin: coinIn)
-                }, onError: { [weak self] error in
-                    self?.stopWaitingApprove()
-                    self?.allowanceStateRelay.accept(.failed(error))
-
-                    self?.sync()
-                })
-
-        allowanceDisposable?.disposed(by: disposeBag)
-    }
-
-    private func updateBalances() {
-        self.balanceIn = self.balance(coin: coinIn)
-        self.balanceOut = self.balance(coin: coinOut)
-    }
-
-    private func sync() {
-        guard let coinIn = coinIn else {        // not select coin in. Just idle
-            stateRelay.accept(.idle)
-            return
-        }
-
-        guard let balanceIn = balanceIn else {      // can't get balance for coinIn. Idle and N/A for balanceIn
-            validationErrorsRelay.accept([SwapValidationError.unavailableBalance(type: .exactIn)])
-            stateRelay.accept(.idle)
-            return
-        }
-
-        var errors = [Error]()
-
-        if let allowanceError = allowanceStateRelay.value?.error {  // can't get allowance
-            errors.append(allowanceError)
-        }
-
-        if (amount(for: .exactIn) ?? 0) > balanceIn {   // not enough balance for swap
-            errors.append(SwapValidationError.insufficientBalance)
-        }
-
-        if let fee = feeState?.data,
-           fee.coinAmount.coin == coinIn,
-           let amount = amount(for: .exactIn),
-           amount + fee.coinAmount.value > balanceIn {  // not enough balance+fee (for Eth) for swap
-
-            errors.append(FeeModule.FeeError.insufficientFeeBalance(coinValue: fee.coinAmount))
-        }
-        if let feeError = feeState?.error {     // can't get fee for swap
-            errors.append(feeError)
-        }
-
-        let hasErrors = !errors.isEmpty
-        if insufficientAllowance() {     // not enough allowance for swap
-            errors.append(SwapValidationError.insufficientAllowance)
-        }
-
-        if hasErrors {
-            validationErrorsRelay.accept(errors)
-            stateRelay.accept(.idle)
-            return
-        }
-
-        if approvingTimer != nil {
-            validationErrorsRelay.accept(errors)
-            stateRelay.accept(.waitingForApprove)
-            return
-        }
-
-        var state = SwapModule.SwapState.idle
-        let forbiddenTrade = tradeDataState?.data?.priceImpactLevel == .forbidden
-
-        if tradeDataState?.data != nil,
-           feeState?.data != nil,
-           !forbiddenTrade {
-
-            state = .proceedAllowed
-        }
-
-        if let allowanceData = allowanceStateRelay.value {
-            let needed = allowanceAmount(for: estimated) ?? 0
-
-            switch allowanceData {
-            case .completed(let allowance): state = needed > allowance && !forbiddenTrade ? .approveRequired : state
-            default: state = .idle
-            }
-        }
-
-        validationErrorsRelay.accept(errors)
-
-        guard stateRelay.value != state else {
-            return
-        }
-        stateRelay.accept(state)
-    }
-
 }
 
 extension SwapService {
 
-    private func handle(tradeData: TradeData, coinIn: Coin, coinOut: Coin) {
-        self.tradeData = tradeData
-        updateFeeState()
-
-        let estimatedAmount = tradeData.type == .exactIn ? tradeData.amountOut : tradeData.amountIn
-        switch tradeData.type {
-        case .exactIn:
-            amountOut = estimatedAmount
-        case .exactOut:
-            amountIn = estimatedAmount
-        }
-
-        var impactLevel = SwapModule.PriceImpactLevel.none
-        if let priceImpact = tradeData.priceImpact {
-            switch priceImpact {
-            case 0..<SwapService.warningPriceImpact: impactLevel = .normal
-            case SwapService.warningPriceImpact..<SwapService.forbiddenPriceImpact: impactLevel = .warning
-            default: impactLevel = .forbidden
-            }
-        }
-
-        let tradeItem = SwapModule.TradeItem(
-                coinIn: coinIn,
-                coinOut: coinOut,
-                type: tradeData.type,
-                executionPrice: tradeData.executionPrice,
-                priceImpact: tradeData.priceImpact,
-                priceImpactLevel: impactLevel,
-                providerFee: tradeData.providerFee,
-                minMaxAmount: tradeData.type == .exactIn ? tradeData.amountOutMin : tradeData.amountInMax)
-
-        tradeDataState = .completed(tradeItem)
-
-        sync()
+    var stateObservable: Observable<State> {
+        stateRelay.asObservable()
     }
 
-    private func handle(allowance: Decimal, coin: Coin) {
-        if let lastAllowance = lastAllowance,
-           allowance == lastAllowance {
-
-            let newState: DataStatus<Decimal> = stateRelay.value == .waitingForApprove ? .loading : .completed(allowance)
-            allowanceStateRelay.accept(newState)
-            return
-        }
-
-        allowanceStateRelay.accept(.completed(allowance))
-        stopWaitingApprove()
-
-        self.lastAllowance = allowance
-        sync()
+    var swapEventObservable: Observable<SwapEvent> {
+        swapEventRelay.asObservable()
     }
 
-}
-
-extension SwapService {
-
-    func tokensForSelection(type: TradeType) -> [SwapModule.CoinBalanceItem] {
-        switch type {
-        case .exactIn: return swapCoinProvider.coins(accountCoins: true, exclude: [])
-        case .exactOut: return swapCoinProvider.coins(accountCoins: false, exclude: [])
-        }
-    }
-
-    func onChange(type: TradeType, amount: Decimal?) {
-        estimated = type
-
-        guard self.amount(for: type) != amount else {
-            return
-        }
-
-        clearEstimated(for: type)
-
-        switch type {
-        case .exactIn:
-            amountIn = amount
-        case .exactOut:
-            amountOut = amount
-        }
-
-        stopWaitingApprove()
-        feeState = nil
-        updateTradeData()
-
-        sync()
-    }
-
-    func onSelect(type: TradeType, coin: Coin) {
-        guard self.coin(for: type) != coin else {
-            return
-        }
-
-        switch type {
-        case .exactIn:
-            coinIn = coin
-
-            if coinOut == coin {
-                coinOut = nil
-            }
-        case .exactOut:
-            coinOut = coin
-
-            if coinIn == coin {
-                coinIn = nil
-            }
-        }
-        updateBalances()
-
-        clearEstimated(for: estimated)
-
-        stopWaitingApprove()
-        feeState = nil
-        updateTradeData()
-
-        sync()
-    }
-
-    func didApprove() {
-        startWaitingApprove()
-
-        sync()
-    }
-
-    func proceed() {
-    }
-
-    func switchCoins() {
-        let swapCoin = coinIn
-        coinIn = coinOut
-        coinOut = swapCoin
-
-        updateBalances()
-
-        clearEstimated(for: estimated)
-
-        stopWaitingApprove()
-        feeState = nil
-        updateTradeData()
-
-        sync()
-    }
-
-    func swap() {
-        guard let tradeData = tradeData,
-              let feeData = feeState?.data else {
-            return
-        }
-
-        swapStateRelay.accept(.loading)
-        swapDisposable?.dispose()
-
-        self.swapStateRelay.accept(.completed(Data()))
-        self.stateRelay.accept(.swapSuccess)
-//        swapDisposable = uniswapRepository
-//                .swap(tradeData: tradeData, gasLimit: feeData.gasLimit, gasPrice: feeData.gasPrice)
-//                .subscribe(onSuccess: { [weak self] _ in
-//                    self?.swapStateRelay.accept(.completed(Data()))
-//
-//                    self?.stateRelay.accept(.swapSuccess)
-//                }, onError: { [weak self] error in
-//                    self?.swapStateRelay.accept(.failed(error))
-//                })
-
-        swapDisposable?.disposed(by: disposeBag)
-    }
-
-    var approveData: SwapModule.ApproveData? {
-        guard let coinIn = coinIn, let amount = allowanceAmount(for: estimated) else {
-            return nil
-        }
-        let allowance = allowanceStateRelay.value?.data ?? 0
-
-        return SwapModule.ApproveData(coin: coinIn,
-                spenderAddress: uniswapRepository.routerAddress,
-                amount: BigUInt(amount.roundedString(decimal: coinIn.decimal)) ?? 0,
-                allowance: BigUInt(allowance.roundedString(decimal: coinIn.decimal)) ?? 0)
-    }
-
-    var estimatedObservable: Observable<TradeType> {
-        estimatedRelay.asObservable()
-    }
-
-    var coinInObservable: Observable<Coin?> {
-        coinInRelay.asObservable()
-    }
-
-    var coinOutObservable: Observable<Coin?> {
-        coinOutRelay.asObservable()
-    }
-
-    var amountInObservable: Observable<Decimal?> {
-        amountInRelay.asObservable()
-    }
-
-    var amountOutObservable: Observable<Decimal?> {
-        amountOutRelay.asObservable()
+    var errorsObservable: Observable<[Error]> {
+        errorsRelay.asObservable()
     }
 
     var balanceInObservable: Observable<Decimal?> {
@@ -541,28 +250,62 @@ extension SwapService {
         balanceOutRelay.asObservable()
     }
 
-    var validationErrorsObservable: Observable<[Error]> {
-        validationErrorsRelay.asObservable()
+    var approveData: SwapAllowanceService.ApproveData? {
+        guard let amount = tradeService.amountIn else {
+            return nil
+        }
+
+        return allowanceService.approveData(amount: amount)
     }
 
-    var allowanceObservable: Observable<DataStatus<Decimal>?> {
-        allowanceStateRelay.asObservable()
+    func swap() {
+        guard case .ready = state, let transaction = transactionService.transactionStatus.data else {
+            return
+        }
+
+        swapEventRelay.accept(.swapping)
+
+        ethereumKit.sendSingle(
+                        address: transaction.data.to,
+                        value: transaction.data.value,
+                        transactionInput: transaction.data.input,
+                        gasPrice: transaction.gasData.gasPrice,
+                        gasLimit: transaction.gasData.gasLimit
+                )
+                .subscribeOn(ConcurrentDispatchQueueScheduler(qos: .background))
+                .observeOn(MainScheduler.instance)
+                .subscribe(onSuccess: { [weak self] transactionWithInternal in
+                    self?.swapEventRelay.accept(.completed)
+                }, onError: { [weak self] error in
+                    self?.swapEventRelay.accept(.failed(error: error))
+                })
+                .disposed(by: disposeBag)
     }
 
-    var feeStateObservable: Observable<DataStatus<SwapModule.SwapFeeInfo>?> {
-        feeStateRelay.asObservable()
+}
+
+extension SwapService {
+
+    enum State {
+        case loading
+        case ready
+        case notReady
     }
 
-    var tradeDataObservable: Observable<DataStatus<SwapModule.TradeItem>?> {
-        tradeDataStateRelay.asObservable()
+    enum SwapEvent {
+        case swapping
+        case completed
+        case failed(error: Error)
     }
 
-    var stateObservable: Observable<SwapModule.SwapState> {
-        stateRelay.asObservable()
+    enum SwapError: Error {
+        case insufficientBalanceIn
+        case insufficientAllowance
+        case forbiddenPriceImpactLevel
     }
 
-    var swapStateObservable: Observable<DataStatus<Data>?> {
-        swapStateRelay.asObservable()
+    enum TransactionError: Error {
+        case insufficientBalance(requiredBalance: BigUInt)
     }
 
 }
